+++ api/server.js
@@ -1,0 +1,18 @@
+const express = require('express');
+
+const authRouter = require('../auth/auth-router.js');
+const usersRouter = require('../users/users-router.js');
+const db = require('../database/dbConfig.js')
+
+const server = express();
+
+server.use(express.json());
+
+server.use('/api/auth', authRouter);
+server.use('/api/users', usersRouter);
+
+server.get('/', (req, res) => {
+  res.json({ api: 'up' });
+});
+
+module.exports = server;
